requirements.txt
numpy==2.4.2
pyyaml==6.0.3
scikit-learn==1.8.0
xgboost==3.2.0
